Backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
alembic
pydantic-settings
python-jose
passlib[bcrypt]
python-multipart
python-dotenv